session: migrate test-infra to testify for temporary table tests in session_test
